feat(Pipelines): enable soft delete of pipeline (#663)
* feat(Pipelines): enable soft deletion of pipeline

* fix(Pipelines): update permission and constraint

* feat(Pipelines): create SoftDeletableMixin

* feat(Core): fine-tune SoftDeleteModel

* Update hexa/core/tests/soft_delete/tests.py

* fix(core): fix failing soft_delete tests

---------

+++ hexa/pipelines/models.py
@@ -25,6 +25,12 @@ from hexa.core.models import (
 )
 from hexa.core.models.base import BaseQuerySet
 from hexa.core.models.behaviors import Status
+from hexa.core.models.soft_delete import (
+    DefaultSoftDeletedManager,
+    IncludeSoftDeletedManager,
+    SoftDeletedModel,
+    SoftDeleteQuerySet,
+)
 from hexa.user_management.models import User
 from hexa.workspaces.models import Workspace, WorkspaceMembership
 
@@ -147,20 +153,23 @@ class PipelineVersion(models.Model):
         return self.display_name
 
 
-class PipelineQuerySet(BaseQuerySet):
+class PipelineQuerySet(BaseQuerySet, SoftDeleteQuerySet):
     def filter_for_user(self, user: AnonymousUser | User):
         return self._filter_for_user_and_query_object(
-            user, Q(workspace__members=user), return_all_if_superuser=False
+            user,
+            Q(workspace__members=user),
+            return_all_if_superuser=False,
         )
 
 
-class Pipeline(models.Model):
+class Pipeline(SoftDeletedModel):
     class Meta:
         verbose_name = "Pipeline"
         constraints = [
             models.UniqueConstraint(
                 "workspace_id",
                 "code",
+                "deleted_at",
                 name="unique_pipeline_code_per_workspace",
             )
         ]
@@ -183,7 +192,8 @@ class Pipeline(models.Model):
     memory_limit = models.CharField(blank=True, max_length=32)
     recipients = models.ManyToManyField(User, through="PipelineRecipient")
 
-    objects = PipelineQuerySet.as_manager()
+    objects = DefaultSoftDeletedManager.from_queryset(PipelineQuerySet)()
+    all_objects = IncludeSoftDeletedManager.from_queryset(PipelineQuerySet)()
 
     def run(
         self,
@@ -257,6 +267,12 @@ class Pipeline(models.Model):
     def delete_if_has_perm(self, *, principal: User):
         if not principal.has_perm("pipelines.delete_pipeline", self):
             raise PermissionDenied
+
+        if PipelineRun.objects.filter(
+            pipeline=self, state__in=[PipelineRunState.QUEUED, PipelineRunState.RUNNING]
+        ).exists():
+            raise PermissionDenied
+
         self.delete()
 
     @property
